4 mars 2018

+++ New Unity Project/Assets/Script/Stats/StatsScript.cs
@@ -9,8 +9,7 @@ public class StatsScript: ScriptableObject {
     public int maxOverHealth = 200;
     public int maxShield = 200;
     [Space]
-    public int maxSpeed = 400;
+    public float maxSpeed = 400.0f;
     public int maxPhysicalDamage = 10;
     public int maxMagicalDamage = 20;
-
 }
